Add a CommentingVisitor
Adds comments to statements if there is an attribute on the node set at some other point. This works around having to manipulate parents to get comments. Next iteration would be to remove the need for knowing a parent and setting an attribute on the childnodes 'parent statement' or something similar.